+++ test/src/models/form_group_test.dart
@@ -538,7 +538,7 @@ void main() {
       form.focus('name');
 
       // Then: control is focused
-      expect((form.control('name') as FormControl).focused, true,
+      expect((form.control('name') as FormControl).hasFocus, true,
           reason: 'control is not focused');
     });
 
@@ -554,8 +554,29 @@ void main() {
       form.focus('person.name');
 
       // Then: control is focused
-      expect((form.control('person.name') as FormControl).focused, true,
+      expect((form.control('person.name') as FormControl).hasFocus, true,
           reason: 'control is not focused');
+    });
+
+    test('Remove Focus to all control', () {
+      // Given: a group
+      final form = FormGroup({
+        'name': FormControl<String>(),
+        'email': FormControl<String>(),
+      });
+
+      // And: all control with focus
+      form.focus('name');
+      form.focus('email');
+
+      // When: remove focus to a control
+      form.unfocus();
+
+      // Then: any control has focus
+      expect((form.control('name') as FormControl).hasFocus, false,
+          reason: 'control is focused');
+      expect((form.control('email') as FormControl).hasFocus, false,
+          reason: 'control is focused');
     });
   });
 }