+++ CHANGELOG.md
@@ -1,6 +1,25 @@
+
+# 5.0.3
+
+## Breaking changes
+- Rename **FormControl.focused** by **FormControl.hasFocus**.
+- Change **ReactiveFormBuilder.builder(context)** by **ReactiveFormBuilder.builder()**.
+- Change argument in FormControl.focusChanges event from *bool* to **FocusEvent**.
+
+## Features
+- Add **FormGroup.unfocus()** and **FormArray.unfocus()** to remove focus of children controls.
+- Add **FormArray.focus(String name)** to set focus on a control.
+- Add optional argument to **AbstractControl.unfocus(bool touched)** to mark controls as untouched 
+when remove focus.
+- Add optional argument to **AbstractControl.reset(bool removeFocus)** to remove focus on control
+when reset the control/form/array.
+
+## Fixes
+- Fix **ReactiveFormBuilder** initializations in debug mode.
+
 # 5.0.2
 
--Fix add custom valueAccessor to text field.
+- Fix add custom valueAccessor to text field.
 
 # 5.0.1